+++ src/full/Agda/Main.hs
@@ -198,7 +198,7 @@ runAgdaWithOptions interactor progName opts = do
             Bench.print
 
             -- Print accumulated statistics.
-            printStatistics 1 Nothing =<< useTC lensAccumStatistics
+            printStatistics Nothing =<< useTC lensAccumStatistics
   where
     -- Options are fleshed out here so that (most) errors like
     -- "bad library path" are validated within the interactor,