+++ CHANGELOG.md
@@ -126,6 +126,12 @@ Library management
   files for `A.B.C` have to reside in `Root`, or further up the
   directory hierarchy.
 
+Pragmas and options
+-------------------
+
+* Profiling options are now turned on with a new `--profile` flag instead of
+  abusing the debug verbosity option. (See [#5781](https://github.com/agda/agda/issues/5731).)
+
 Performance
 -----------